requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
JPype1==1.5.0
mariadb==1.1.10
MarkupSafe==2.1.5
packaging==24.1
python-dotenv==1.0.1
Werkzeug==3.0.3
